test(curl): keep path tests within repository layout
